+++ sys/main.c
@@ -67,7 +67,7 @@ void start(uint32_t *modulep, void *physbase, void *physfree)
   kprintf("physfree %p\n", (uint64_t)physfree);
   kprintf("tarfs in [%p:%p]\n", &_binary_tarfs_start, &_binary_tarfs_end);
   task_struct *pcb_boot = (task_struct *)kmalloc(sizeof(task_struct));  //kernel 
-  pcb_boot->pml4 =(uint64_t)ker_pml4_t;  // kernel's page table   
+  pcb_boot->pml4 = ker_pml4_t;  // kernel's page table   
   pcb_boot->cr3 = ker_cr3; // kernel's page table   
   pcb_boot->pid = PID++;  // I'm kernel init process  so pid 0
